Floor division uses integer division for integral arguments (#12131)
Previously (since f0ac8dae2) floor division for all pairs of arguments
was implemented by promoting both arguments to double, dividing, and
then taking the floor.

This produces results when both operands are integral that sometimes
don't match the behaviour of normal floor division in any of the
target languages that cuDF supports. Specifically for int64 operands,
any time an operand is larger than $2^{53}$, the conversion to double is
not injective and so information is lost.

To address this, implement SFINAE overloads for the FloorDiv device
functor that use standard C++ type promotion semantics for division
with mixed type operands. In the case of integral types with same
sign, this is then just standard division; if the signs are mixed, we
round towards negative infinity; if either type is floating, we
perform floating point division and then take the floor.

These operations are now explicitly not implemented for Decimal and
duration dtypes (rather than implicitly not implemented in the wrapper
layers).

If the calling library's requirements for type promotion differ from
those of C++, it is the now the responsibility of the caller to
promote the operands to floor division before calling into libcudf.

Closes #12120.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Yunsong Wang (https://github.com/PointKernel)
  - Bradley Dice (https://github.com/bdice)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/12131