Pin arrow 9 in testing dependencies to prevent conda solve issues (#12377)
This pinning is necessary because other testing dependencies require arrow (currently, transformers has a transitive dependency on it) and don't include a pinning as strict as ours, resulting in a newer arrow package (arrow 10) being installed. When we attempt to install our cudf/libcudf packages in CI, it should be able to downgrade to arrow 9, but the downgrade fails because changes in arrow's package structure in arrow 10 include the creation of a new libarrow package (arrow-cpp was renamed to libarrow) and libarrow has no equivalent in arrow 9. As a result, conda does not see a way to downgrade it, and it also does not see a way to satisfy its dependencies if the rest of the arrow packages are downgraded to arrow 9 (because libarrow is tied to arrow 10 packages), so the conda solve fails.

The current changeset will fix Github Actions CI. We may need additional fixes in our Jenkins CI.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/12377