Add optional text file support to ninja-log utility (#12823)
Adds support for `sort_ninja_log.py` utility to accept an optional text file to be included at the top of the generated html report. This allows it to be used asynchronously outside the build step and helps enable using it in C++ builds in non-cudf repos.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/12823