Fix parquet schema interpretation issue (#13277)
There is a bug reading parquet files that have a specific encoding. The typical list is 
```
<list-repetition> group <name> (LIST) {
  repeated group list {
    <element-repetition> <element-type> element;
  }
}
```
but it can also be
```
<list-repetition> group <name> (LIST) {
  repeated group list {
    repeated int32 name
  }
}
```
This second case was failing and this fixes that. The issue was the `one_level_list` was returning true for the name, which resulted in another nesting level. The fix is to only return true if the immediate parent is not a list.

closes #13237
closes #13239

Authors:
  - Mike Wilson (https://github.com/hyperbolic2346)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Yunsong Wang (https://github.com/PointKernel)
  - https://github.com/nvdbaranec
  - MithunR (https://github.com/mythrocks)
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/cudf/pull/13277