Add benchmark (Google Benchmark) dependency to conda packages. (#13528)
While working with @DaceT on benchmarks, we found that the benchmark executables in the `libcudf-tests` conda package are not usable because the `libcudf-tests` package is missing a dependency on `benchmark` (Google Benchmark). This PR adds that dependency.

Currently Google Benchmark is fetched by rapids-cmake CPM, and is pinned to version `1.8.0`. This PR matches that pinning, and ensures that a consistent package is used at build time and run time.

This PR also adds a minimal benchmark to CI tests to ensure that benchmarks are runnable.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Yunsong Wang (https://github.com/PointKernel)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/13528