Fixes a performance regression in FST (#13850)
https://github.com/rapidsai/cudf/pull/13344 introduced a performance regression to the FST benchmarks that showed as much as a 35% performance degradation. 

It seems that, after the refactor in the above PR, compiler optimization heuristics are deciding differently on loop unrolling in the part of the FST that's writing out transduced symbols. 

As a fix, we are enforcing to not unroll that loop.

Authors:
  - Elias Stehle (https://github.com/elstehle)

Approvers:
  - Karthikeyan (https://github.com/karthikeyann)
  - David Wendt (https://github.com/davidwendt)

URL: https://github.com/rapidsai/cudf/pull/13850