Add noSanitizer tag to Java reduction tests failing with sanitizer in CUDA 12 (#13904)
Relates to NVIDIA/spark-rapids-jni#1349.  The Java ReductionTest unit tests are failing when run under CUDA 12's compute-sanitizer but pass when run with the CUDA 11 version.  To unblock CI, marking the affected tests to be run without the sanitizer in the interim while this is being investigated.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Gera Shegalov (https://github.com/gerashegalov)

URL: https://github.com/rapidsai/cudf/pull/13904